sprint-2/step-1 - Правки согласно замечаниям
