+++ src/compontents/BurgerIngredients/BurgerIngredients.jsx
@@ -8,20 +8,19 @@ import ingredientsTypes from '../../utils/types'
 
 // @ts-ignore
 const BurgerIngredients = ({ ingredients }) => {
-
+    const [currentTab, setCurrentTab] = React.useState('one')
 
     function ingredientsTabs() {
-        // eslint-disable-next-line react-hooks/rules-of-hooks
-        const [current, setCurrent] = React.useState('one')
+
         return (
             <div style={{display: 'flex'}}>
-                <Tab value="one" active={current === 'one'} onClick={setCurrent}>
+                <Tab value="one" active={currentTab === 'one'} onClick={setCurrentTab}>
                     Булки
                 </Tab>
-                <Tab value="two" active={current === 'two'} onClick={setCurrent}>
+                <Tab value="two" active={currentTab === 'two'} onClick={setCurrentTab}>
                     Соусы
                 </Tab>
-                <Tab value="three" active={current === 'three'} onClick={setCurrent}>
+                <Tab value="three" active={currentTab === 'three'} onClick={setCurrentTab}>
                     Начинки
                 </Tab>
             </div>
